+++ CHANGES.rst
@@ -5,8 +5,10 @@ Changelog
 1.0.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Add some default metadata in summary serializer.
+  [cekk]
+- Return getObjSize and mime_type in summary serializer for brains and contents  with enhancedlinks behavior.
+  [cekk]
 
 1.0.0 (2024-01-26)
 ------------------